[NRT-508] Move update_notes_with_named_undo helper to main/utils
The helper wraps aqt.mw.col.update_notes in a labeled custom undo entry.
It's collection-level (not GUI-specific), so it fits better alongside the
other aqt-aware collection helpers in main/utils.py than under gui/utils.
